requirements.txt
httpx==0.27.0
pillow==10.4.0
pyperclip==1.9.0
pynput==1.7.7
pygetwindow==0.0.9
openai==1.37.0
pyqt6==6.6.1
pyqt6-qt6==6.6.1
pyqt6-sip==13.6.0